[#16] feat: :sparkles: 수량정보 컴포넌트 분리
