+++ src/component/popUp/PopUpInformations.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Label from '../util/Label';
-import { FaAngleUp, FaAngleDown } from 'react-icons/fa';
+import PopUpItemCountContainer from './PopUpItemCountContainer';
 
 const PopUpInformationsStyle = styled.div`
 `;
@@ -57,44 +57,6 @@ const PopUpItemBuyingInformationTitle = styled.div`
 
 const PopUpItemBuyingInformationContent = styled.div`
     line-height: 1.5rem;
-`;
-const PopUpItemCountContainer = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 1.5rem 0;
-    border-bottom: 1px solid #e0e0e0;
-    & > *:last-child {
-        display: grid;
-        grid-template-columns: 5rem 2rem;
-        line-height: 0;
-    }
-`;
-
-const PopUpItemCountTitle = styled.div`
-    color: #828282;
-    width: 5rem;
-`;
-
-const PopUpItemCount = styled.input.attrs({
-  type: 'text'
-})`
-    font-size: 1rem;
-    padding: 0rem 1rem;
-    border: 1px solid #E0E0E0;
-    text-align: center;
-`;
-
-const PopUpItemCountBtns = styled.div`
-    border: 1px solid #E0E0E0;
-    border-left: 0px;
-`;
-
-const PopUpItemCountBtn = styled.button`
-    width: 100%;
-    &:first-child {
-        border-bottom: 1px solid #E0E0E0;
-    }
 `;
 
 const PopUpItemTotalPriceContainer = styled.div`
@@ -159,16 +121,7 @@ export default function PopUpInformations({ detailData }) {
                     <PopUpItemBuyingInformationContent>{delivery_fee}</PopUpItemBuyingInformationContent>
                 </PopUpItemBuyingInformation>
             </PopUpItemBuyingInformations>
-            <PopUpItemCountContainer>
-                <PopUpItemCountTitle>수량</PopUpItemCountTitle>
-                <div>
-                    <PopUpItemCount></PopUpItemCount>
-                    <PopUpItemCountBtns>
-                        <PopUpItemCountBtn><FaAngleUp/></PopUpItemCountBtn>
-                        <PopUpItemCountBtn><FaAngleDown/></PopUpItemCountBtn>
-                    </PopUpItemCountBtns>
-                </div>
-            </PopUpItemCountContainer>
+            <PopUpItemCountContainer stock={stock}/>
             <PopUpItemTotalPriceContainer>
                 <PopUpItemTotalPriceTitle>총 주문금액</PopUpItemTotalPriceTitle>
                 <PopUpTotalPrice>{price}</PopUpTotalPrice>
